Change the flow of the payment with QuickPay
A generated orderId and the QuickPay paymentId are now used as
Shopware transactionId and paymentUniqueId. Writing the order
to the database was moved to the successAction to prevent premature
ending of the the ordering process. Once a payment id is received
from Quickpay it is stored in the session until the order is finished.
Identification of the order in the callbackAction is accordingly done
by using the order_id and id of the received payment object.